Add time to http request log
This commit adds code to prefix lines in the request log with the current time and date.

JavaScript seemingly has no native support for ISO8601 (i.e. the only sensible date format) which honors the server's timezone settings (what the hell, JavaScript?) so the date and time are instead formatted using the Swedish locale, which is a very near approximation.  Have also adopted this stategy for date formatting to replace the somewhat cumbersome construction previously used to generate game IDs.